Changes:
Game now displays pieces as string objects on the board.

A players pieces can only be moved if it is the players turn currently, the players turn updates accordingly after a piece is moved.

Games are saved to files and can be loaded from the file. Duplicates currently allowed (this will be changed).